+++ src/getDataFromTree.ts
@@ -53,6 +53,8 @@ export function walkTree<Cache>(
       // In case the user doesn't pass these to super in the constructor
       instance.props = instance.props || props;
       instance.context = instance.context || context;
+      // set the instance state to null (not undefined) if not set, to match React behaviour
+      instance.state = instance.state || null;
 
       // Override setState to just change the state, not queue up an update.
       //   (we can't do the default React thing as we aren't mounted "properly"