+++ Changelog.md
@@ -2,10 +2,11 @@
 
 ### vNext
 - BREAKING: removed cleanupApolloState as it is no longer needed!
-- Exported getDataFromTree on the client 
+- Exported getDataFromTree on the client
 - Removed `redux` from peer dependencies. [Issue #1223](https://github.com/apollographql/react-apollo/issues/1223) [PR #1224](https://github.com/apollographql/react-apollo/pull/1224)
 - Support arrays being returned from render in SSR [#1158](https://github.com/apollographql/react-apollo/pull/1158)
 - Support passing an updater function to `setState` in SSR mode [#1263](https://github.com/apollographql/react-apollo/pull/1263)
+- Correctly initializes component state as null (not undefined) [#1300](https://github.com/apollographql/react-apollo/pull/1310)
 
 ### 2.0.0-beta.0
 - upgrade to Apollo Client 2.0